requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
SQLAlchemy>=1.4.0,<2.0.0
PyMySQL>=1.1.0,<1.2.0
lightfm>=1.16,<2.0
python-dotenv>=1.0.0,<1.1.0